In-memory Cache for front-end server
Cache2k is a lightweight and fast in-memory caching library that
pygmy.com uses to reduce network requests.

This commit adds the cache2k jars and a basic implementation
of the topic/book lookup cache.

Verified that JSON responses for a topc/book-id are cached and
subsequent requests for the same topic/book-id are server from
the cache.