Support load balancing in UIServer
Implemented a simple round-robin load balancer
for UIServer.

The LoadBalancer is essentially a BlockingQueue
to which servers are queued. Every search/lookup/buy request
gets forwarded to the server at head of the queue, then the
head is removed and that server is added to the tail
of the queue. BlockingQueue was chosen because its concurrent
thread safe.