Implemented state machine replication for CatalogServer
We use a distributed  token-ring algorithm to ensure
that only one replcia executed an order at a given time.
The replica that has the token executes an order, then
replicates the same on on the other replica, releases
the token to the replica and only then acknowledges
completion of task.